Change how we root all application libraries
When `rootallapplicationassemblies` is specified, we currently go over all assemblies passed to ILC as input or reference, find out which ones are not framework assemblies, and root all types in them so that they're available for arbitrary reflection at runtime.

The problem is that NuGet packages or SDKs can have many assemblies in them, but only some of them are actually used by user code. Good example is the Microsoft.WindowsDesktop.App SDK that has both WinForms and WPF in it.

With this change, we'll only root assemblies where we actually saw the app using something in them. This way a WinForms app will not include all of WPF.

We could possibly regress apps that do `Assembly.Load` on an assembly that is otherwise unused. Those apps will have to use RD.XML. We could also add a simple data flow analysis for the simplest cases to avoid RD.XML.

ASP.NET still works without RD.XML for this.

Most of this commit is just mechanically moving stuff around.
