+++ src/ILCompiler.Compiler/src/Compiler/RootingServiceProvider.cs
@@ -1,0 +1,123 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using ILCompiler.DependencyAnalysis;
+
+using Internal.TypeSystem;
+
+using Debug = System.Diagnostics.Debug;
+
+namespace ILCompiler
+{
+    internal delegate void RootAdder(object o, string reason);
+
+    internal class RootingServiceProvider : IRootingServiceProvider
+    {
+        private readonly NodeFactory _factory;
+        private readonly RootAdder _rootAdder;
+
+        public RootingServiceProvider(NodeFactory factory, RootAdder rootAdder)
+        {
+            _factory = factory;
+            _rootAdder = rootAdder;
+        }
+
+        public void AddCompilationRoot(MethodDesc method, string reason, string exportName = null)
+        {
+            MethodDesc canonMethod = method.GetCanonMethodTarget(CanonicalFormKind.Specific);
+            IMethodNode methodEntryPoint = _factory.MethodEntrypoint(canonMethod);
+            _rootAdder(methodEntryPoint, reason);
+
+            if (exportName != null)
+                _factory.NodeAliases.Add(methodEntryPoint, exportName);
+
+            if (canonMethod != method && method.HasInstantiation)
+                _rootAdder(_factory.MethodGenericDictionary(method), reason);
+        }
+
+        public void AddCompilationRoot(TypeDesc type, string reason)
+        {
+            _rootAdder(_factory.MaximallyConstructableType(type), reason);
+        }
+
+        public void RootThreadStaticBaseForType(TypeDesc type, string reason)
+        {
+            Debug.Assert(!type.IsGenericDefinition);
+
+            MetadataType metadataType = type as MetadataType;
+            if (metadataType != null && metadataType.ThreadGcStaticFieldSize.AsInt > 0)
+            {
+                _rootAdder(_factory.TypeThreadStaticIndex(metadataType), reason);
+
+                // Also explicitly root the non-gc base if we have a lazy cctor
+                if (_factory.TypeSystemContext.HasLazyStaticConstructor(type))
+                    _rootAdder(_factory.TypeNonGCStaticsSymbol(metadataType), reason);
+            }
+        }
+
+        public void RootGCStaticBaseForType(TypeDesc type, string reason)
+        {
+            Debug.Assert(!type.IsGenericDefinition);
+
+            MetadataType metadataType = type as MetadataType;
+            if (metadataType != null && metadataType.GCStaticFieldSize.AsInt > 0)
+            {
+                _rootAdder(_factory.TypeGCStaticsSymbol(metadataType), reason);
+
+                // Also explicitly root the non-gc base if we have a lazy cctor
+                if (_factory.TypeSystemContext.HasLazyStaticConstructor(type))
+                    _rootAdder(_factory.TypeNonGCStaticsSymbol(metadataType), reason);
+            }
+        }
+
+        public void RootNonGCStaticBaseForType(TypeDesc type, string reason)
+        {
+            Debug.Assert(!type.IsGenericDefinition);
+
+            MetadataType metadataType = type as MetadataType;
+            if (metadataType != null && (metadataType.NonGCStaticFieldSize.AsInt > 0 || _factory.TypeSystemContext.HasLazyStaticConstructor(type)))
+            {
+                _rootAdder(_factory.TypeNonGCStaticsSymbol(metadataType), reason);
+            }
+        }
+
+        public void RootVirtualMethodForReflection(MethodDesc method, string reason)
+        {
+            Debug.Assert(method.IsVirtual);
+
+            if (method.HasInstantiation)
+            {
+                _rootAdder(_factory.GVMDependencies(method), reason);
+            }
+            else
+            {
+                // Virtual method use is tracked on the slot defining method only.
+                MethodDesc slotDefiningMethod = MetadataVirtualMethodAlgorithm.FindSlotDefiningMethodForVirtualMethod(method);
+                if (!_factory.VTable(slotDefiningMethod.OwningType).HasFixedSlots)
+                    _rootAdder(_factory.VirtualMethodUse(slotDefiningMethod), reason);
+            }
+
+            if (method.IsAbstract)
+            {
+                _rootAdder(_factory.ReflectableMethod(method), reason);
+            }
+        }
+
+        public void RootModuleMetadata(ModuleDesc module, string reason)
+        {
+            // RootModuleMetadata is kind of a hack - this is pretty much only used to force include
+            // type forwarders from assemblies metadata generator would normally not look at.
+            // This will go away when the temporary RD.XML parser goes away.
+            if (_factory.MetadataManager is UsageBasedMetadataManager)
+                _rootAdder(_factory.ModuleMetadata(module), reason);
+        }
+
+        public void RootReadOnlyDataBlob(byte[] data, int alignment, string reason, string exportName)
+        {
+            var blob = _factory.ReadOnlyDataBlob("__readonlydata_" + exportName, data, alignment);
+            _rootAdder(blob, reason);
+            _factory.NodeAliases.Add(blob, exportName);
+        }
+    }
+}
