+++ src/ILCompiler/src/ILCompiler.csproj
@@ -39,7 +39,6 @@
     <Compile Include="Program.cs" />
     <Compile Include="RdXmlRootProvider.cs" />
     <Compile Include="EcmaOnlyDebugInformationProvider.cs" />
-    <Compile Include="ApplicationAssemblyRootProvider.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\..\Common\src\CommandLine\CommandLineException.cs">
