+++ src/ILCompiler/src/RdXmlRootProvider.cs
@@ -72,7 +72,7 @@ namespace ILCompiler
 
                 foreach (TypeDesc type in ((EcmaModule)assembly).GetAllTypes())
                 {
-                    RootType(rootProvider, type, "RD.XML root");
+                    RootingHelpers.TryRootType(rootProvider, type, "RD.XML root");
                 }
             }
 
@@ -103,7 +103,7 @@ namespace ILCompiler
                 if (dynamicDegreeAttribute.Value != "Required All")
                     throw new NotSupportedException();
 
-                RootType(rootProvider, type, "RD.XML root");
+                RootingHelpers.RootType(rootProvider, type, "RD.XML root");
             }
 
             foreach (var element in typeElement.Elements())
@@ -150,85 +150,7 @@ namespace ILCompiler
                 method = method.MakeInstantiatedMethod(methodInst);
             }
 
-            RootMethod(rootProvider, method, "RD.XML root");
-        }
-
-        public static void RootType(IRootingServiceProvider rootProvider, TypeDesc type, string reason)
-        {
-            rootProvider.AddCompilationRoot(type, reason);
-
-            // Instantiate generic types over something that will be useful at runtime
-            if (type.IsGenericDefinition)
-            {
-                Instantiation inst = TypeExtensions.GetInstantiationThatMeetsConstraints(type.Instantiation, allowCanon: true);
-                if (inst.IsNull)
-                    return;
-
-                type = ((MetadataType)type).MakeInstantiatedType(inst);
-
-                rootProvider.AddCompilationRoot(type, reason);
-            }
-
-            // Also root base types. This is so that we make methods on the base types callable.
-            // This helps in cases like "class Foo : Bar<int> { }" where we discover new
-            // generic instantiations.
-            TypeDesc baseType = type.BaseType;
-            while (baseType != null)
-            {
-                baseType = baseType.NormalizeInstantiation();
-                RootType(rootProvider, baseType, reason);
-                baseType = baseType.BaseType;
-            }
-
-            if (type.IsDefType)
-            {
-                foreach (var method in type.GetMethods())
-                {
-                    if (method.HasInstantiation)
-                    {
-                        // Generic methods on generic types could end up as Foo<object>.Bar<__Canon>(),
-                        // so for simplicity, we just don't handle them right now to make this more
-                        // predictable.
-                        if (!method.OwningType.HasInstantiation)
-                        {
-                            Instantiation inst = TypeExtensions.GetInstantiationThatMeetsConstraints(method.Instantiation, allowCanon: false);
-                            if (!inst.IsNull)
-                            {
-                                RootMethod(rootProvider, method.MakeInstantiatedMethod(inst), reason);
-                            }
-                        }
-                    }
-                    else
-                    {
-                        RootMethod(rootProvider, method, reason);
-                    }
-                }
-            }
-        }
-
-        private static void RootMethod(IRootingServiceProvider rootProvider, MethodDesc method, string reason)
-        {
-            try
-            {
-                LibraryRootProvider.CheckCanGenerateMethod(method);
-
-                // Virtual methods should be rooted as if they were called virtually
-                if (method.IsVirtual)
-                    rootProvider.RootVirtualMethodForReflection(method, reason);
-
-                if (!method.IsAbstract)
-                    rootProvider.AddCompilationRoot(method, reason);
-            }
-            catch (TypeSystemException)
-            {
-                // TODO: fail compilation if a switch was passed
-
-                // Individual methods can fail to load types referenced in their signatures.
-                // Skip them in library mode since they're not going to be callable.
-                return;
-
-                // TODO: Log as a warning
-            }
+            RootingHelpers.TryRootMethod(rootProvider, method, "RD.XML root");
         }
     }
 }
